requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.1.0
newrelic==4.6.0.106
PyExecJS==1.5.1
pymongo==3.5.0
raven==6.10.0
requests==2.18.4
six==1.12.0
soupsieve==1.8
urllib3==1.22
Werkzeug==0.14.1
flask-cors==3.0.7
